made eth token values more accurate
previously for each symbol we use the highest marketcap id, this is mostly for other tokens, since for other tokens we are likely not to include trash coins. However, for eth address, there might be a lot of trash coins, so using the highest market cap might get wrong price.

Solution is to provide eth tokens a price overwrite function, which is directly getting back from eth api, to make sure this price is accurate. Furthermore, sometimes the price is undefined, in this case don't overwrite the existing price. So added a overwritePrices() to do this thing.